# TradingBot.Api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "TradingBot API",
    "version": "v1"
  },
  "paths": {
    "/api/Candles": {
      "get": {
        "tags": [
          "Candles"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Candle"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Candle"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Candle"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Candles"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Candle"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Candle"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Candle"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Candle"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candle"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candle"
                }
              }
            }
          }
        }
      }
    },
    "/api/Candles/{id}": {
      "get": {
        "tags": [
          "Candles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Candle"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candle"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Candle"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Candles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Candle"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Candle"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Candle"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Candles"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Candle": {
        "required": [
          "close",
          "createDate",
          "high",
          "low",
          "number",
          "open",
          "timestamp",
          "volume",
          "volumeWeight"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createDate": {
            "type": "string",
            "format": "date-time"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64"
          },
          "open": {
            "type": "number",
            "format": "double"
          },
          "high": {
            "type": "number",
            "format": "double"
          },
          "low": {
            "type": "number",
            "format": "double"
          },
          "close": {
            "type": "number",
            "format": "double"
          },
          "volume": {
            "type": "integer",
            "format": "int64"
          },
          "volumeWeight": {
            "type": "integer",
            "format": "int64"
          },
          "number": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      }
    }
  }
}